+++ packages/plugin-sales-api/src/graphql/resolvers/mutations/utils.ts
@@ -1090,14 +1090,24 @@ export const checkLoyalties = async (
   }
 
   for (const item of activeProductsData || []) {
+    
+    if (item.discountPercent) {
+      continue;
+    }
+
     const loyalty = loyalties[item.productId];
 
     item.unitPrice = item.unitPrice || 0;
 
     if (loyalty) {
+
+      if (!loyalty.discount) {
+        continue;
+      }
+
       item.discountPercent = loyalty.discount
       item.discount = fixNum(((item.quantity * item.unitPrice) / 100) * loyalty.discount);
-      item.amount = fixNum(item.unitPrice - (item.unitPrice / 100) * loyalty.discount);
+      item.amount = fixNum((item.unitPrice - (item.unitPrice / 100) * loyalty.discount) * item.quantity);
     }
   }